moving injection hook stuff into core
