+++ workers/unity/Assets/Playground/Scripts/OneTimeInitialisation.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using Improbable.Gdk.Core;
 using Improbable.Gdk.PlayerLifecycle;
 using Unity.Entities;
 using UnityEngine;
@@ -7,7 +7,7 @@ namespace Playground
 {
     public static class OneTimeInitialisation
     {
-        private static bool initialized = false;
+        private static bool initialized;
 
         [RuntimeInitializeOnLoadMethod]
         private static void Init()
@@ -18,38 +18,11 @@ namespace Playground
             }
 
             initialized = true;
-            SetupInjectionHooks();
-            PlayerLoopManager.RegisterDomainUnload(DomainUnloadShutdown, 1000);
+            CoreInitialization.SetupInjectionHooks();
+            PlayerLoopManager.RegisterDomainUnload(CoreInitialization.DomainUnloadShutdown, 1000);
 
             // Setup template to use for player on connecting client
             PlayerLifecycleConfig.CreatePlayerEntityTemplate = PlayerTemplate.CreatePlayerEntityTemplate;
-        }
-
-        private static void SetupInjectionHooks()
-        {
-            var hybridAssembly = typeof(GameObjectEntity).Assembly;
-
-            // Reflection to get internal hook classes. Doesn't seem to be a proper way to do this.
-            var gameObjectArrayInjectionHookType =
-                hybridAssembly.GetType("Unity.Entities.GameObjectArrayInjectionHook");
-            var transformAccessArrayInjectionHookType =
-                hybridAssembly.GetType(
-                    "Unity.Entities.TransformAccessArrayInjectionHook");
-            var componentArrayInjectionHookType =
-                hybridAssembly.GetType("Unity.Entities.ComponentArrayInjectionHook");
-
-            InjectionHookSupport.RegisterHook(
-                (InjectionHook) Activator.CreateInstance(gameObjectArrayInjectionHookType));
-            InjectionHookSupport.RegisterHook(
-                (InjectionHook) Activator.CreateInstance(transformAccessArrayInjectionHookType));
-            InjectionHookSupport.RegisterHook(
-                (InjectionHook) Activator.CreateInstance(componentArrayInjectionHookType));
-        }
-
-        private static void DomainUnloadShutdown()
-        {
-            World.DisposeAllWorlds();
-            ScriptBehaviourUpdateOrder.UpdatePlayerLoop();
         }
     }
 }
