+++ workers/unity/Packages/com.improbable.gdk.core/CoreInitialization.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4202796876e1b449a8c939f66094258
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
